Shell linting - improve robustness of variables in shell scripts fdf2dcm.sh:  new error_exit function, ID and version updated

+++ fdf2dcm.sh
@@ -5,23 +5,32 @@
 # - Michael Eager (michael.eager@monash.edu.au)
 # - Monash Biomedical Imaging 
 # - (C) 2014 Michael Eager
-
+#
+#
+#  "$Id:"
+#  Version 0.0: Wrapper for agilent2dicom
+#  Version 1.0: Full support of FDF formats
+###############################################
 
 set -o nounset  # shortform: -u
 set -o errexit  # -e
 # set -o pipefail
-#touch $(dirname $0)/error.log
+# touch $(dirname $0)/error.log
 #exec 2>> $(dirname $0)/error.log
-# set -x
+#set -x  # show debugging output
 
 
-# Set config variables
-VERBOSE=0
-MODIFY=1
+## Set config variables
+FDF2DCMVERSION=1.1
+PROGNAME=$(basename $0)
 FDF2DCMPATH=$(dirname $0)
 KERNEL_RELEASE=$(uname -r | awk -F'.' '{printf("%d.%d.%d\n", $1,$2,$3)}')
 DCM3TOOLS="${FDF2DCMPATH}/../dicom3tools_1.00.snapshot.20140306142442/bin/1.${KERNEL_RELEASE}.x8664/"
-# Check DCMTK on MASSIVE or Agilent console
+## Set dcmulti and its arguments
+DCMULTI="dcmulti -v -makestack -sortby AcquisitionNumber -dimension StackID FrameContentSequence -dimension InStackPositionNumber FrameContentSequence -of "
+DCMULTI_DTI="dcmulti -v -makestack -sortby DiffusionBValue -dimension StackID FrameContentSequence -dimension InStackPositionNumber FrameContentSequence -of "
+#-makestack -sortby ImagePositionPatient  -sortby AcquisitionNumber
+# Check dcmtk applications on MASSIVE or Agilent console
 if test ${MASSIVEUSERNAME+defined}; then
     test -x dcmodify || module load dcmtk 
 else
@@ -36,12 +45,14 @@ elif [ ! -x ${DCM3TOOLS}/dcmulti ]; then
     exit 1
 fi 
 export PATH=${PATH}:${DCM3TOOLS}
-input_dir=""
-output_dir=""
-python_args=""
-
+declare -i verbosity=0
+declare -i do_modify=1
+declare input_dir=""
+declare output_dir=""
+declare python_args=""
 
 E_BADARGS=65
+
 #source ${FDF2DCMPATH}/yesno.sh
 function yesno(){
     read -r -p "$@" response
@@ -52,11 +63,10 @@ function yesno(){
     fi
     return 1
 }
-
-DCMULTI="dcmulti -v -makestack -sortby AcquisitionNumber -dimension StackID FrameContentSequence -dimension InStackPositionNumber FrameContentSequence -of "
-DCMULTI_DTI="dcmulti -v -makestack -sortby DiffusionBValue -dimension StackID FrameContentSequence -dimension InStackPositionNumber FrameContentSequence -of "
-#-makestack -sortby ImagePositionPatient  -sortby AcquisitionNumber
-
+function error_exit(){
+    echo "${PROGNAME}: ${1:-Unknown error}" 1>&2
+    exit 1
+}
 
 
 # Print usage information and exit
@@ -66,15 +76,16 @@ print_usage(){
 	"\n" \
 	"-i <inputdir>  FDF source directory\n" \
 	"-o <outputdir> Optional destination DICOM directory. Default is input_dir/.dcm. \n" \
-	"-v             verbose output. \n" \
+	"-d             Disable dcmodify fixes to DICOMs." \
 	"-m,-p          Enable magnitude and phase subdirectory conversion.  These flags are passed to agilent2dicom and should only be used from within fdf2dcm or with knowledge of input fdf data. \n" \
+	"-v             verbose output. \n" \
 	"-h             this help\n" \
 	"\n" 
     # && exit 1
 }
 
 
-# Check for number of args
+## Check for number of args
 if [ $# -eq 0 ]; then
     echo "fdfdcm.sh must have one argument: -i, --input [directory of FDF images]"
     print_usage
@@ -82,9 +93,7 @@ if [ $# -eq 0 ]; then
 fi
 
 
-
-
-# Parge arguments
+## Parse arguments
 while getopts ":i:o:hmpdv" opt; do
     case $opt in
 	i)
@@ -109,12 +118,12 @@ while getopts ":i:o:hmpdv" opt; do
 	    python_args="$python_args -p"
 	    ;;
 	d)
-	    MODIFY=0
+	    do_modify=0
 	    echo " Disable dcmodify correction."
 	    ;;
 	v)
-	    echo "Setting verbose to 1."
-	    VERBOSE=1
+	    verbosity=1
+	    echo "Setting verbose to $verbosity."
 	    python_args="$python_args -v"
 	    ;;
 	\?)
@@ -136,27 +145,28 @@ if [ ! -d "$input_dir" ]; then
     echo "fdfdcm.sh must have a valid input directory of FDF images."
     exit $E_BADARGS
 fi
-# Set output_dir if not in args, default is INPUT/.dcm
+## Set output_dir if not in args, default is INPUT/.dcm
 if [ "$output_dir" == "" ]; then
     output_dir="$(dirname ${input_dir})/$(basename ${input_dir} .img).dcm"
     echo "Output dir set to: " $output_dir
 fi
 
-
+## Check for MAG/PHASE component directories
 if [ -d  ${input_dir}/magnitude.img ] || [ -d ${input_dir}/phase.img ]; then
     echo "Input directory has 'magnitude.img' and 'phase.img' "
-    verb=''; if [ "$VERBOSE" -eq 1 ]; then verb=' -v '; fi 
+    verb=''; if [ "$verbosity" -eq 1 ]; then verb=' -v '; fi 
     $0 $verb -m -i ${input_dir}/magnitude.img/ -o ${output_dir}/magnitude.dcm/
-    if [ "$VERBOSE" -eq 1 ] && ! yesno "Magnitude complete. Continue converting phase?"; then
+    if [ "$verbosity" -eq 1 ] && ! yesno "Magnitude complete. Continue converting phase?"; then
 	echo "fdf2dcm completed without phase." && exit 0
     fi	
     $0 $verb -p -i ${input_dir}/phase.img/ -o ${output_dir}/phase.dcm/
     exit 0
 fi
 
+## Check output directory
 JumpToDCmulti=0
 if [ -d "${output_dir}" ]; then
-    if [ -d "${output_dir}/tmp" -a "$VERBOSE" -eq 1 ];then
+    if [ -d "${output_dir}/tmp" -a "$verbosity" -eq 1 ];then
 	if yesno "Remove existing output directory, y or n (default y)?"; then
 	    echo "Removing existing output directory"
 	    rm -rf ${output_dir}
@@ -169,47 +179,41 @@ if [ -d "${output_dir}" ]; then
     fi	
 fi
 
-if [ $JumpToDCmulti -eq 0 ]; then
-
-# Carefull with nullglob on csh  
-    shopt -s nullglob
+if (( "${JumpToDCmulti}" == 0 ))
+then
+    shopt -s nullglob  # Carefull with nullglob on csh  
     found=0
     for i in ${input_dir}/*.fdf; do
 	((++found))
     done
     shopt -u nullglob
     if [ $found -eq 0 ]; then  #-o "$fdffiles" == "" 
-	echo "fdf2dcm Error: Input directory has no FDF images"
-	exit 1
+	error_exit "$LINENO: Input directory has no FDF images"
     else
 	echo $found, " FDF files were found"
     fi
 
     if [ ! -f ${input_dir}/procpar ]; then
-	echo "fdf2dcm Error: Input directory has no procpar file"
-	exit 1
+	error_exit "$LINENO: Input directory has no procpar file"
     fi
 
 # set -o errexit  # -e
 # set -o pipefail
 
 
-## Crux of script
+## Crux of script - conversion of FDF images to standard DICOM images
     echo  "Calling agilent2dicom"
     echo " Arguments: ", ${python_args} -i "${input_dir}" -o "${output_dir}"
     ${FDF2DCMPATH}/agilent2dicom.py ${python_args} -i "${input_dir}" -o "${output_dir}"
 
-    if [ $? -ne 0 ]; then 
-	echo "Error: agilent2dicom failed"
-	exit 1
-    fi
-    [ ! -d ${output_dir} ] && (echo "Output dir not created by agilent2dicom. Shutting down fdf2dcm."; exit 1)
+    [ $? -ne 0 ] && error_exit "$LINENO: agilent2dicom failed"
+    
+    [ ! -d ${output_dir} ] && error_exit "$LINENO: Output dir not created by agilent2dicom."
 
-    dcmfiles=`ls ${output_dir}/*.dcm`
-    if [ $? -ne 0 ]; then  #-o "$fdffiles" == "" 
-	echo "Error: Output directory of agilent2dicom has no DICOM images. Shutting down fdf2dcm."
-	exit 1
-    fi
+    # dcmfiles=$(ls ${output_dir}/*.dcm)  ## Bad code - use glob
+    #if[ $? -ne 0 ]
+    test -e ${output_dir}/0001.dcm  &&	error_exit "$LINENO: Output directory of agilent2dicom has no DICOM images."
+    
 
     echo "Moving dicom images"
     mkdir ${output_dir}/tmp
@@ -255,37 +259,43 @@ echo "DCMULTI complete. Fixing inconsistencies."
 
 
 ## Corrections to dcmulti conversion
-${FDF2DCMPATH}/fix-dicoms.sh "${output_dir}"
-echo "Fixing dicoms complete."
+if [ ${do_modify} -eq 1 ]
+then
+    ${FDF2DCMPATH}/fix-dicoms.sh "${output_dir}"
+    echo "Fixing dicoms complete."
 ## Additional corrections to diffusion files
-if [ -f ${output_dir}/DIFFUSION ];then
-    ${FDF2DCMPATH}/fix-diffusion.sh "${output_dir}"
-    echo "Fixed diffusion module parameters."
-    rm -f ${output_dir}/DIFFUSION
+    if [ -f ${output_dir}/DIFFUSION ];then
+	${FDF2DCMPATH}/fix-diffusion.sh "${output_dir}"
+	echo "Fixed diffusion module parameters."
+	rm -f ${output_dir}/DIFFUSION
+    fi
 fi
+[ -f ${output_dir}/DIFFUSION ] && rm -f ${output_dir}/DIFFUSION
 
-
-if [ "$VERBOSE" -eq 1 ]; then
+if [ "$verbosity" -eq 1 ]; then
     echo "Verifying dicom compliance using dciodvfy."
-    if [ -f "$output_dir"/0001.dcm ]; then
-	dciodvfy $output_dir/0001.dcm &> $(dirname $output_dir)/$(basename $output_dir .dcm).log
+    if [ -f "${output_dir}"/0001.dcm ]; then
+	## Send dciodvfy stderr and stdout to log file
+	dciodvfy ${output_dir}/0001.dcm &> $(dirname ${output_dir})/$(basename ${output_dir} .dcm).log  
     else
-	echo "ERROR: could not find $output_dir/0001.dcm for verification"
+	error_exit "$LINENO: could not find ${output_dir}/0001.dcm for verification"
     fi
 fi
 
-## Cleaning up
+## Cleaning up temporary directories
 if [ -d ${output_dir}/tmp ]; then
-    if [ "$VERBOSE" -eq 1 ]; then
+    if [ "$verbosity" -eq 1 ]; then
 	if yesno "Remove existing tmp output directory, y or n (default y)?"; then
 	    echo "Removing existing tmp output directory"
 	    rm -rf ${output_dir}/tmp    
 	else
 	    echo "fdf2dcm completed. Temporary dicoms still remain."
+	    exit 0
 	fi
     else
 	echo "Removing existing tmp output directory"
 	rm -rf ${output_dir}/tmp    
     fi
+    [ -d ${output_dir}/tmp ] && error_exit "$LINENO: temporary dicom directory could not be deleted."
 fi
